backend/pyproject.toml
[project]
name = "urban-tree-observatory"
version = "0.1.0"
description = "Urban Tree Observatory"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "django-allauth[mfa,socialaccount]>=65.5.0",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-leaflet>=0.31.0",
    "django-storages[s3]>=1.14.5",
    "djangorestframework>=3.15.2",
    "djangorestframework-gis>=1.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "gunicorn>=23.0.0",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "psycopg>=3.2.6",
    "sentry-sdk>=2.24.0",
    "tqdm>=4.67.1",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.1.0",
    "django-extensions>=3.2.3",
    "factory-boy>=3.3.3",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-django>=4.10.0",
    "ruff>=0.11.2",
]
